requirements.txt
pandas
evaluate
absl-py==2.1.0
accelerate==0.29.3
aiohttp
aiosignal
annotated-types==0.6.0
anthropic==0.29.0
anyio==4.3.0
argcomplete==1.12.3
ase==3.21.1
asgiref==3.8.1
asttokens
async-timeout
attrs
backcall
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
boto==2.49.0
build==1.2.1
cached-property==1.5.2
cachetools==5.3.3
certifi
cffi==1.14.6
chardet==4.0.0
charset-normalizer
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpickle==1.6.0
colorama
coloredlogs==15.0.1
coolname==2.2.0
crcmod==1.7
cryptography
cycler==0.10.0
dataclasses-json==0.6.6
datasets==2.19.1
decorator==4.4.2
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
eval_type_backport==0.2.0
evaluate==0.4.2
exceptiongroup
executing
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.16.3
filelock==3.15.4
FLAML==2.1.2
flatbuffers==24.3.25
frozenlist
fsspec
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.19.1
httptools==0.6.1
httpx==0.27.0
huggingface-hub>=0.34.0
humanfriendly==10.0
idna==2.10
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython
isodate==0.6.0
jedi
Jinja2
jiter==0.5.0
joblib==1.0.1
jsonpatch==1.33
jsonpointer==2.4
klepto==0.2.5
kiwisolver==1.3.1
kubernetes==30.1.0
langchain==0.1.5
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe
marshmallow==3.21.2
matplotlib
matplotlib-inline
mdurl==0.1.2
mkl-fft
mkl-random
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
mpmath
multidict
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx>=3.2
nltk==3.6.2
numba==0.59.1
numpy>=1.23.0,<2.0
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.11.1
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==1.5.3
parso
peft==0.10.0
pexpect
pickleshare
pillow>=10.0.0
posthog==3.5.0
pox==0.2.9
prompt-toolkit
protobuf==4.25.3
psutil
ptyprocess
pure-eval
pyarrow
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==2.7.0
pydantic_core==2.18.1
pydot==1.4.2
pyglet==1.5.0
Pygments
pyOpenSSL
pyparsing==2.4.7
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks
python-dateutil==2.8.1
python-dotenv==1.0.1
python-louvain==0.15
python-multipart==0.0.9
python-rapidjson==1.16
pytz==2021.1
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==23.2.1
rdflib==5.0.0
redis==3.5.3
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
retry-decorator==1.1.1
rich==13.7.1
rouge_score==0.1.2
rsa==4.7.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.0.2
scipy>=1.11.0
seaborn==0.13.0
sentence-transformers
sh==1.14.3
shellingham==1.5.4
six==1.15.0
smart-open==5.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data
starlette==0.37.2
sympy
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.4.0
threadpoolctl==2.1.0
tiktoken==0.7.0
together==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch>=2.0.0
torchaudio
torchvision
tqdm==4.66.4
traitlets
transformers>=4.30.0
triton
tritonclient==2.33.0
typer==0.12.3
typing-inspect==0.6.0
typing_extensions
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.1.2
wrapt==1.16.0
xxhash==3.4.1
yarl
zipp==3.18.1
zope.event==5.0
zope.interface==6.3 
